+++ app/client/src/actions/pageActions.tsx
@@ -26,7 +26,7 @@ import type { UrlDataState } from "reducers/entityReducers/appReducer";
 import type { APP_MODE } from "entities/App";
 import type { CanvasWidgetsReduxState } from "reducers/entityReducers/canvasWidgetsReducer";
 import type { GenerateTemplatePageRequest } from "api/PageApi";
-import type { ENTITY_TYPE } from "entities/AppsmithConsole";
+import type { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 import type { Replayable } from "entities/Replay/ReplayEntity/ReplayEditor";
 import * as Sentry from "@sentry/react";
