+++ app/client/src/ce/components/editorComponents/Debugger/entityTypeLinkMap.tsx
@@ -2,7 +2,7 @@ import ActionLink from "components/editorComponents/Debugger/ActionLink";
 import DatasourceLink from "components/editorComponents/Debugger/DataSourceLink";
 import WidgetLink from "components/editorComponents/Debugger/WidgetLink";
 import JSCollectionLink from "components/editorComponents/Debugger/JSCollectionLink";
-import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 
 export const entityTypeLinkMap = {
   [ENTITY_TYPE.WIDGET]: WidgetLink,
