+++ app/client/src/components/editorComponents/CodeEditor/codeEditorUtils.ts
@@ -1,5 +1,5 @@
 import type CodeMirror from "codemirror";
-import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 import type {
   WidgetEntity,
   ActionEntity,
