+++ app/client/src/components/editorComponents/Debugger/ActionLink.tsx
@@ -1,5 +1,5 @@
 import { PluginType } from "entities/Action";
-import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 import React, { useCallback } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import type { AppState } from "@appsmith/reducers";
