+++ app/client/src/components/editorComponents/Debugger/hooks/debuggerHooks.ts
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 import { useParams } from "react-router";
 import type { Log } from "entities/AppsmithConsole";
-import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 import type { AppState } from "@appsmith/reducers";
 import { getWidget } from "sagas/selectors";
 import {
