+++ app/client/src/components/editorComponents/JSResponseView.tsx
@@ -50,7 +50,7 @@ import {
 import LogHelper from "./Debugger/ErrorLogs/components/LogHelper";
 import LOG_TYPE from "entities/AppsmithConsole/logtype";
 import type { SourceEntity, Log } from "entities/AppsmithConsole";
-import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 import { CloseDebugger } from "./Debugger/DebuggerTabs";
 
 const ResponseContainer = styled.div`
