+++ app/client/src/ee/entities/AppsmithConsole/utils.ts
@@ -1,0 +1,1 @@
+export * from "ce/entities/AppsmithConsole/utils";
