+++ app/client/src/entities/AppsmithConsole/index.ts
@@ -3,18 +3,10 @@ import type LOG_TYPE from "./logtype";
 import type { PropertyEvaluationErrorType } from "utils/DynamicBindingUtils";
 import type { PluginType } from "entities/Action";
 import type { HTTP_METHOD } from "constants/ApiEditorConstants/CommonApiConstants";
-
-export enum ENTITY_TYPE {
-  ACTION = "ACTION",
-  DATASOURCE = "DATASOURCE",
-  WIDGET = "WIDGET",
-  JSACTION = "JSACTION",
-}
-
-export enum PLATFORM_ERROR {
-  PLUGIN_EXECUTION = "PLUGIN_EXECUTION",
-  JS_FUNCTION_EXECUTION = "JS_FUNCTION_EXECUTION",
-}
+import type {
+  ENTITY_TYPE,
+  PLATFORM_ERROR,
+} from "@appsmith/entities/AppsmithConsole/utils";
 
 export type Methods =
   | "log"
