+++ app/client/src/entities/AppsmithConsole/logtype.ts
@@ -13,6 +13,7 @@ enum LOG_TYPE {
   JS_PARSE_SUCCESS,
   CYCLIC_DEPENDENCY_ERROR,
   LINT_ERROR,
+  MISSING_MODULE,
 }
 
 export default LOG_TYPE;
