+++ app/client/src/entities/Replay/ReplayEntity/ReplayCanvas.ts
@@ -11,7 +11,7 @@ import {
   WIDGETS,
 } from "../replayUtils";
 import type { AppTheme } from "entities/AppTheming";
-import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 
 export interface Canvas {
   widgets: CanvasWidgetsReduxState;
