+++ app/client/src/pages/Editor/DataSourceEditor/RestAPIDatasourceForm.tsx
@@ -36,7 +36,7 @@ import _ from "lodash";
 import FormLabel from "components/editorComponents/FormLabel";
 import CopyToClipBoard from "components/designSystems/appsmith/CopyToClipBoard";
 import { updateReplayEntity } from "actions/pageActions";
-import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 import { TEMP_DATASOURCE_ID } from "constants/Datasource";
 import { Form } from "./DBForm";
 import { selectFeatureFlagCheck } from "@appsmith/selectors/featureFlagsSelectors";
