+++ app/client/src/sagas/OneClickBindingSaga.ts
@@ -44,7 +44,7 @@ import ActionAPI from "api/ActionAPI";
 import { validateResponse } from "./ErrorSagas";
 import AnalyticsUtil from "utils/AnalyticsUtil";
 import AppsmithConsole from "utils/AppsmithConsole";
-import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 import { fetchActions, runAction } from "actions/pluginActionActions";
 import { Toaster, Variant } from "design-system-old";
 import WidgetFactory from "WidgetProvider/factory";
