+++ app/client/src/sagas/QueryPaneSagas.ts
@@ -58,7 +58,7 @@ import {
   startFormEvaluations,
 } from "actions/evaluationActions";
 import { updateReplayEntity } from "actions/pageActions";
-import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 import type { EventLocation } from "@appsmith/utils/analyticsUtilTypes";
 import AnalyticsUtil from "utils/AnalyticsUtil";
 import {
