+++ app/client/src/selectors/debuggerSelectors.test.ts
@@ -1,4 +1,5 @@
-import { Severity, ENTITY_TYPE, LOG_CATEGORY } from "entities/AppsmithConsole";
+import { Severity, LOG_CATEGORY } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "@appsmith/entities/AppsmithConsole/utils";
 import LOG_TYPE from "entities/AppsmithConsole/logtype";
 import type { DataTree } from "entities/DataTree/dataTreeTypes";
 import type { CanvasWidgetsReduxState } from "reducers/entityReducers/canvasWidgetsReducer";
